tests/wpt/web-platform-tests/css/css-fonts/parsing/font-variant-position-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: parsing font-variant-position with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-variant-position-prop">
<meta name="assert" content="font-variant-position supports only the grammar 'normal | sub | super'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value('font-variant-position', 'auto');
test_invalid_value('font-variant-position', 'super sub');
</script>
</body>
</html>
